feat(auth): keep scoped_service_jwt importable before the app registry
Modules loaded at Django startup declare purposes at module level, so the
purpose half must not import models or DRF. The DRF authentication class
moves to posthog.auth beside the other authentication classes.

Generated-By: PostHog Desktop
Task-Id: 8071b078-abb1-46b0-b9ea-aa6b96b38d9b
