+++ posthog/auth.py
@@ -6,7 +6,7 @@ import logging
 import functools
 from abc import abstractmethod
 from datetime import timedelta
-from typing import TYPE_CHECKING, Any, Optional, TypedDict, Union
+from typing import TYPE_CHECKING, Any, ClassVar, Optional, TypedDict, Union
 from urllib.parse import parse_qs, urlparse
 
 from django.apps import apps
@@ -56,6 +56,7 @@ from posthog.models.utils import (
 from posthog.models.webauthn_credential import WebauthnCredential
 from posthog.passkey import verify_passkey_authentication_response
 from posthog.rbac.user_access_control import UserAccessControl
+from posthog.scoped_service_jwt import ScopedServiceJwtPurpose
 from posthog.shared_link_user import SharedLinkUser
 from posthog.synthetic_user import SyntheticUser
 
@@ -1080,6 +1081,90 @@ class InternalAPIAuthentication(authentication.BaseAuthentication):
 
     def authenticate_header(self, request: HttpRequest) -> str:
         return self.keyword
+
+
+class ScopedServiceJWTAuthentication(authentication.BaseAuthentication):
+    """DRF authentication for internal routes called by other PostHog services, verifying the
+    scoped per-purpose JWTs of posthog.scoped_service_jwt (the blessed replacement for
+    INTERNAL_API_SECRET). Subclass per purpose:
+
+        class ConversationsTicketJWTAuthentication(ScopedServiceJWTAuthentication):
+            purpose = CONVERSATIONS_TICKETS_PURPOSE
+
+    The request authenticates as a synthetic InternalAPIUser bound to the token's team.
+    When the URL carries a team_id, the token's team_id claim must match it, so a token
+    minted for one team can never read another's data even if both hit the same route.
+
+    require_team=False is for fleet-scoped purposes (cron-style calls with no team in the
+    URL or claims); team-scoped purposes must keep the default so a token without a team
+    claim fails closed instead of authenticating unscoped.
+    """
+
+    purpose: ClassVar[ScopedServiceJwtPurpose]
+    require_team: ClassVar[bool] = True
+
+    def authenticate(self, request: Request) -> Optional[tuple[Any, Any]]:
+        header = authentication.get_authorization_header(request).split()
+        # No bearer header: return None (not raise) so the view's other authenticators,
+        # if any, still get their turn.
+        if not header or header[0].lower() != b"bearer":
+            return None
+        if len(header) != 2:
+            raise AuthenticationFailed("Invalid Authorization header.")
+
+        if not self.purpose.enabled():
+            # Fail closed at request time rather than startup: most processes never serve
+            # these routes and never get the secret injected, so a startup check would
+            # wrongly crash them (same reasoning as InternalAPIAuthentication).
+            logger.error(
+                "Scoped service JWT authentication attempted without configured secret",
+                extra={"path": request.path, "settings_name": self.purpose.settings_name},
+            )
+            raise AuthenticationFailed("Service token authentication is not configured.")
+
+        try:
+            claims = self.purpose.verify(header[1].decode())
+        except jwt.PyJWTError:
+            raise AuthenticationFailed("Invalid or expired service token.")
+
+        return self._authenticate_claims(request, claims)
+
+    def _authenticate_claims(self, request: Request, claims: dict[str, Any]) -> tuple[Any, Any]:
+        claim_team_id = claims.get("team_id")
+
+        if claim_team_id is None:
+            if self.require_team:
+                raise AuthenticationFailed("Service token is missing its team claim.")
+            return InternalAPIUser(), None
+
+        url_team_id = _team_id_from_request_path(request)
+        if url_team_id is not None and str(claim_team_id) != url_team_id:
+            raise AuthenticationFailed("Service token team does not match the requested team.")
+
+        Team = apps.get_model(app_label="posthog", model_name="Team")
+        try:
+            team = Team.objects.only("id", "organization_id").get(id=claim_team_id)
+        except (Team.DoesNotExist, ValueError, TypeError):
+            raise AuthenticationFailed("Invalid service token team.")
+
+        return InternalAPIUser(current_organization_id=team.organization_id, current_team_id=team.id), None
+
+
+def _team_id_from_request_path(request: Request) -> Optional[str]:
+    parser_context = getattr(request, "parser_context", None)
+    if isinstance(parser_context, dict):
+        kwargs = parser_context.get("kwargs")
+        if isinstance(kwargs, dict) and kwargs.get("team_id") is not None:
+            return str(kwargs["team_id"])
+
+    django_request = getattr(request, "_request", request)
+    resolver_match = getattr(django_request, "resolver_match", None)
+    if resolver_match and getattr(resolver_match, "kwargs", None):
+        team_id = resolver_match.kwargs.get("team_id")
+        if team_id is not None:
+            return str(team_id)
+
+    return None
 
 
 def session_auth_required(endpoint):
